Update extension.js
I've tried this change in the code to enable the glow on my computer. When I was trying to enable the glow, VS Code send me this message "You must run VS code with admin privileges in order to enable Neon Dreams.". Looking for it on the Ethernet, I found this option proposed by the user @pulberg on https://github.com/robb0wen/synthwave-vscode/issues/263